# backend/artifacts/openapi-v6.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearPath Nexus Engine",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/weather/environmental": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Get Environmental Risks",
        "operationId": "get_environmental_risks_api_v1_weather_environmental_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentalRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weather/map-conditions": {
      "post": {
        "tags": [
          "weather"
        ],
        "summary": "Get Map Conditions",
        "operationId": "get_map_conditions_api_v1_weather_map_conditions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapConditionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapConditionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/weather/route-points": {
      "post": {
        "tags": [
          "weather"
        ],
        "summary": "Get Route Weather Points",
        "operationId": "get_route_weather_points_api_v1_weather_route_points_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteWeatherPointsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteWeatherPointsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/weather/kp-index": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Get Kp Index",
        "operationId": "get_kp_index_api_v1_weather_kp_index_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Kp Index Api V1 Weather Kp Index Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnvironmentalRiskResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "weather": {
            "additionalProperties": true,
            "type": "object",
            "title": "Weather"
          },
          "space_weather": {
            "additionalProperties": true,
            "type": "object",
            "title": "Space Weather"
          },
          "weather_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Score"
          },
          "alerts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "status",
          "source",
          "weather",
          "space_weather",
          "weather_score",
          "alerts"
        ],
        "title": "EnvironmentalRiskResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MapConditionFailure": {
        "properties": {
          "point_id": {
            "type": "string",
            "title": "Point Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "point_id",
          "message"
        ],
        "title": "MapConditionFailure"
      },
      "MapConditionPoint": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "MapConditionPoint"
      },
      "MapConditionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "reading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "category",
          "lat",
          "lon"
        ],
        "title": "MapConditionResponse"
      },
      "MapConditionsRequest": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/MapConditionPoint"
            },
            "type": "array",
            "minItems": 1,
            "title": "Points"
          },
          "destination_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10,
                "pattern": "^[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Code"
          }
        },
        "type": "object",
        "required": [
          "points"
        ],
        "title": "MapConditionsRequest"
      },
      "MapConditionsResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "title": "Status"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/MapConditionResponse"
            },
            "type": "array",
            "title": "Conditions"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "failures": {
            "items": {
              "$ref": "#/components/schemas/MapConditionFailure"
            },
            "type": "array",
            "title": "Failures"
          }
        },
        "type": "object",
        "required": [
          "status",
          "conditions",
          "source"
        ],
        "title": "MapConditionsResponse"
      },
...
      "RouteWeatherPointRequest": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          },
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon",
          "id"
        ],
        "title": "RouteWeatherPointRequest"
      },
      "RouteWeatherPointResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "temperature_2m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature 2M"
          },
          "apparent_temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apparent Temperature"
          },
          "relative_humidity_2m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Humidity 2M"
          },
          "weather_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Code"
          },
          "wind_speed_10m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed 10M"
          },
          "wind_direction_10m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Direction 10M"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          },
          "uv_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uv Index"
          },
          "precipitation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lat",
          "lon",
          "available"
        ],
        "title": "RouteWeatherPointResponse"
      },
      "RouteWeatherPointsRequest": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/RouteWeatherPointRequest"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "points"
        ],
        "title": "RouteWeatherPointsRequest"
      },
      "RouteWeatherPointsResponse": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/RouteWeatherPointResponse"
            },
            "type": "array",
            "title": "Points"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "open-meteo"
          }
        },
        "type": "object",
        "required": [
          "points"
        ],
        "title": "RouteWeatherPointsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}